+++ client/pages/Reports/Reports.tsx
@@ -4,17 +4,16 @@ import List from 'components/List'
 import { IListGroups } from 'components/List/types'
 import { value as value } from 'helpers'
 import { Spinner } from 'office-ui-fabric-react/lib/Spinner'
+import { format } from 'office-ui-fabric-react/lib/Utilities'
 import * as React from 'react'
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { exportExcel } from 'utils/exportExcel'
 import columns from './columns'
+import { exportToExcel, openFilterPanel, selectGroupBy, selectQuery } from './commands'
 import styles from './Reports.module.scss'
 import TIME_ENTRIES, { ITimeEntriesVariables } from './TIME_ENTRIES'
-import { IReportsQuery, getGroupByOptions, getQueries } from './types'
-import { pick } from 'underscore'
-import { IContextualMenuItem } from 'office-ui-fabric-react'
-import { format } from 'office-ui-fabric-react/lib/Utilities'
+import { IReportsQuery } from './types'
 
 /**
  * @category Reports
@@ -33,7 +32,7 @@ export const Reports = () => {
         emptyGroupName: t('all'),
     })
     const [subset, setSubset] = useState<any[]>(undefined)
-    const { loading, error, data } = useQuery<any, ITimeEntriesVariables>(
+    const { loading, data } = useQuery<any, ITimeEntriesVariables>(
         TIME_ENTRIES,
         {
             skip: !query,
@@ -69,9 +68,6 @@ export const Reports = () => {
         setSubset(_entries)
     }
 
-    const queries = getQueries(t)
-    const groupByOptions = getGroupByOptions(t)
-
     return (
         <div className={styles.root}>
             <List
@@ -87,49 +83,13 @@ export const Reports = () => {
                 enableShimmer={loading}
                 commandBar={{
                     items: [
-                        {
-                            key: 'SELECT_QUERY',
-                            text: query?.text || t('selectReportLabel', { ns: 'reports' }),
-                            iconProps: { iconName: query?.iconName || 'ReportDocument' },
-                            subMenuProps: {
-                                items: queries.map(query => ({
-                                    ...pick(query, 'key', 'text'),
-                                    iconProps: { iconName: query.iconName },
-                                    onClick: () => setQuery(query),
-                                })),
-                            }
-                        },
-                        {
-                            key: 'GROUP_BY',
-                            text: t('groupBy'),
-                            iconProps: { iconName: 'GroupList' },
-                            disabled: !query || loading,
-                            subMenuProps: {
-                                items: groupByOptions.map(opt => ({
-                                    ...pick(opt, 'key', 'text'),
-                                    canCheck: true,
-                                    checked: groupBy.fieldName === opt.props.fieldName,
-                                    onClick: () => setGroupBy(opt.props),
-                                } as IContextualMenuItem)),
-                            }
-                        },
-                    ],
+                        selectQuery(query, setQuery, t),
+                        (query && !loading) && selectGroupBy(groupBy, setGroupBy, t),
+                    ].filter(i => i),
                     farItems: [
-                        {
-                            key: 'EXPORT_TO_EXCEL',
-                            text: t('exportCurrentView'),
-                            onClick: onExportExcel,
-                            iconProps: { iconName: 'ExcelDocument' },
-                            disabled: loading || !!error || !query,
-                        },
-                        {
-                            key: 'OPEN_FILTER_PANEL',
-                            iconProps: { iconName: 'Filter' },
-                            iconOnly: true,
-                            onClick: () => setFilterPanelOpen(true),
-                            disabled: loading || !!error || !query,
-                        }
-                    ]
+                        (query && !loading) && exportToExcel(onExportExcel, t),
+                        (query && !loading)  && openFilterPanel(setFilterPanelOpen)
+                    ].filter(i => i)
                 }} />
             {loading && (
                 <Spinner
